feat(fraud-detection): implement event ordering and vector clocks
